Removed Bounce Button, Removed the Android Specific layout, Downgraded to Mono 4.0.30319.17020, Updated Insights Tracking
Removed Bounce Button. Forms was throwing an error because the UI
Animations were overlapping.
Removed the Android Specific layout. The Nexus6P that we will be using
for the demo at Evolve properly aligns the picture based on Horizontal
vs. Vertical.
Downgraded to Mono 4.0.30319.17020. The previous version of Mono (in
Beta release) was throwing runtime errors.
Updated Insights Tracking. Now track if a photo was taken.